+++ watson_developer_cloud/alchemy_language_v1.py
@@ -34,11 +34,21 @@ class AlchemyLanguageV1(WatsonDeveloperCloudService):
         return self._alchemy_html_request('GetAuthors', html=html, url=url, params=params)
 
     def keywords(self, html=None, text=None, url=None, strict_extract_mode=False, sentiment=False,
-                 show_source_text=False, max_keywords=50, language=None):
+                 show_source_text=False, max_items=None, language=None, max_keywords=50):
+        """
+        :param html: HTML input
+        :param text: Text input
+        :param url: URL input
+        :param max_items: The number of results to return (default 50)
+        :param max_keywords: deprecated, use max_items instead
+        :return: A JSON object with extracted keywords from the source document
+        """
+        if not max_items:
+            max_items = max_keywords
         params = {'keywordExtractMode': 'strict' if strict_extract_mode else 'normal',
                   'sentiment': sentiment,
                   'showSourceText': show_source_text,
-                  'maxRetrieve': max_keywords,
+                  'maxRetrieve': max_items,
                   'language': language}
         return self._alchemy_html_request('GetRankedKeywords', html=html, text=text, url=url, params=params)
 
@@ -51,7 +61,7 @@ class AlchemyLanguageV1(WatsonDeveloperCloudService):
         return self._alchemy_html_request('GetRankedConcepts', html=html, text=text, url=url, params=params)
 
     def entities(self, html=None, text=None, url=None, disambiguate=True, linked_data=True, coreference=True,
-                 quotations=False, sentiment=False, show_source_text=False, max_items=50, language=None):
+                 quotations=False, sentiment=False, show_source_text=False, max_items=50, language=None, model=None):
         params = {'disambiguate': disambiguate,
                   'linkedData': linked_data,
                   'coreference': coreference,
@@ -59,7 +69,8 @@ class AlchemyLanguageV1(WatsonDeveloperCloudService):
                   'sentiment': sentiment,
                   'showSourceText': show_source_text,
                   'maxRetrieve': max_items,
-                  'language': language}
+                  'language': language,
+                  'model': model}
         return self._alchemy_html_request('GetRankedNamedEntities', html=html, text=text, url=url, params=params)
 
     def emotion(self, html=None, text=None, url=None, show_source_text=False, source_text_type=None,