+++ watson_developer_cloud/personality_insights_v2.py
@@ -25,14 +25,14 @@ class PersonalityInsightsV2(WatsonDeveloperCloudService):
     """Wrapper of the Personality Insights service"""
     default_url = 'https://gateway.watsonplatform.net/personality-insights/api'
 
-    def __init__(self, url=default_url, username=None, password=None, use_vcap_services=True):
+    def __init__(self, url=default_url, **kwargs):
         """
         Construct an instance. Fetches service parameters from VCAP_SERVICES
         runtime variable for Bluemix, or it defaults to local URLs.
         """
 
         WatsonDeveloperCloudService.__init__(
-            self, 'personality_insights', url, username, password, use_vcap_services)
+            self, 'personality_insights', url, **kwargs)
 
     def profile(self, text, content_type='text/plain', accept='application/json', language=None, csv_headers=False):
         """