+++ watson_developer_cloud/visual_recognition_v3.py
@@ -26,15 +26,15 @@ class VisualRecognitionV3(WatsonDeveloperCloudService):
     default_url = 'https://gateway-a.watsonplatform.net/visual-recognition/api'
     latest_version = '2016-05-20'
 
-    def __init__(self, version, url=default_url, use_vcap_services=True, api_key=None):
+    def __init__(self, version, url=default_url, **kwargs):
         """
         Construct an instance. Fetches service parameters from VCAP_SERVICES
         runtime variable for Bluemix, or it defaults to local URLs.
         :param version: specifies the specific version-date of the service to use, for example '2016-05-20'
+        :param api_key: specifies the credentials for the service (doesn't use username and password)
         """
 
-        WatsonDeveloperCloudService.__init__(
-            self, 'watson_vision_combined', url, None, None, use_vcap_services, api_key)
+        WatsonDeveloperCloudService.__init__(self, 'watson_vision_combined', url, **kwargs)
         self.version = version
 
     def get_classifier(self, classifier_id):
@@ -80,6 +80,20 @@ class VisualRecognitionV3(WatsonDeveloperCloudService):
         # Params sent as url parameters here
         return self.request(method='POST', url='/v3/classifiers', files=kwargs, data=data, params=params,
                             accept_json=True)
+
+    def update_classifier(self, classifier_id, **kwargs):
+        """
+        Updates an existing classifier by adding images to existing or new classes.
+        :param classifier_id: The id of the classifier to update.
+        :param <NAME>_positive_examples: zip files of images that depict the subject of the class.
+        :param negative_examples: A zip file of images that do not depict the subject of any of the classes.
+        :return:
+        """
+
+        params = {'version': self.version}
+        # Params sent as url parameters here
+        return self.request(method='POST', url='/v3/classifiers/{0}'.format(classifier_id), files=kwargs,
+                            params=params, accept_json=True)
 
     def _image_call(self, url, images_file=None, images_url=None, params=None):
         if images_file is None and images_url is None: